update code genre api section

+++ MovieAPI/Controllers/GenreController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Http;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using MovieAPI.Entities;
 using MovieAPI.Services;
@@ -17,12 +18,13 @@ namespace MovieAPI.Controllers
         }
 
         [HttpGet]
-        public IActionResult GetAll()
+        public IActionResult GetAll([FromQuery] QueryObject query)
         {
-            var res = _genreRepository.GetAll();
+            var res = _genreRepository.GetAll(query);
             return Ok(res);
         }
 
+        //[Authorize]
         [HttpPost]
         public IActionResult AddOne(GenreVM genreVM)
         {
@@ -34,17 +36,24 @@ namespace MovieAPI.Controllers
             return Ok(res);
         }
 
+        //[Authorize]
         [HttpDelete("{id}")]
         public IActionResult Delete(Guid id)
         {
+            if(id == Guid.Empty)
+            {
+                return BadRequest();
+            }
+
             var res = _genreRepository.RemoveById(id);
-            if(res == null)
+            if(res.Success == false)
             {
                 return NotFound();
             }
             return Ok(res);
         }
 
+        //[Authorize]
         [HttpPut]
         public IActionResult Update(GenreDTO genre)
         {
@@ -54,7 +63,7 @@ namespace MovieAPI.Controllers
             }
 
             var res = _genreRepository.Update(genre);
-            if(res == null)
+            if(res.Success == false)
             {
                 return NotFound();
             }
